+++ src/pages/player-page/player-page.tsx
